+++ AdventOfCode2019/Day4.cs
@@ -7,20 +7,20 @@ namespace AdventOfCode2019
 {
 	public class Day4
 	{
-		public static int Result1(string input)
+		public static long Result1(string input)
 		{
-			int count = 0;
+			long count = 0;
 
-			int start = int.Parse(input.Split('-')[0]);
-			int end = int.Parse(input.Split('-')[1]);
+			long start = long.Parse(input.Split('-')[0]);
+			long end = long.Parse(input.Split('-')[1]);
 
-			for (int i = start; i <= end; i++)
+			for (long i = start; i <= end; i++)
 			{
 				bool maybe = false;
-				for (int x = 0; x < 5; x++)
+				for (long x = 0; x < 5; x++)
 				{
 					//two of the same digits in a row
-					if (i.ToString()[x] == i.ToString()[x + 1])
+					if (i.ToString()[(int)x] == i.ToString()[(int)x + 1])
 					{
 						maybe = true;
 						break;
@@ -29,11 +29,11 @@ namespace AdventOfCode2019
 
 				if (!maybe) continue;
 				maybe = false;
-				for (int x = 0; x < 5; x++)
+				for (long x = 0; x < 5; x++)
 				{
-					char a = i.ToString()[x];
-					char b = i.ToString()[x + 1];
-					if (i.ToString()[x] > i.ToString()[x + 1])
+					char a = i.ToString()[(int)x];
+					char b = i.ToString()[(int)x + 1];
+					if (i.ToString()[(int)x] > i.ToString()[(int)x + 1])
 					{
 						maybe = false;
 						break;
@@ -53,24 +53,24 @@ namespace AdventOfCode2019
 
 
 
-		public static int Result2(string input)
+		public static long Result2(string input)
 		{
-			int count = 0;
+			long count = 0;
 
-			int start = int.Parse(input.Split('-')[0]);
-			int end = int.Parse(input.Split('-')[1]);
+			long start = long.Parse(input.Split('-')[0]);
+			long end = long.Parse(input.Split('-')[1]);
 
-			List<int> results = new List<int>();
+			List<long> results = new List<long>();
 
-			for (int i = start; i <= end; i++)
+			for (long i = start; i <= end; i++)
 			{
 				bool maybe = false;
 
-				int ct = 0;
-				for (int x = 0; x < 5; x++)
+				long ct = 0;
+				for (long x = 0; x < 5; x++)
 				{
 					//two of the same digits in a row
-					if (i.ToString()[x] == i.ToString()[x + 1])
+					if (i.ToString()[(int)x] == i.ToString()[(int)x + 1])
 					{
 						ct++;
 						maybe = true;
@@ -79,11 +79,11 @@ namespace AdventOfCode2019
 
 				if (!maybe) continue;
 				maybe = false;
-				for (int x = 0; x < 5; x++)
+				for (long x = 0; x < 5; x++)
 				{
-					char a = i.ToString()[x];
-					char b = i.ToString()[x + 1];
-					if (i.ToString()[x] > i.ToString()[x + 1])
+					char a = i.ToString()[(int)x];
+					char b = i.ToString()[(int)x + 1];
+					if (i.ToString()[(int)x] > i.ToString()[(int)x + 1])
 					{
 						maybe = false;
 						break;
@@ -102,16 +102,16 @@ namespace AdventOfCode2019
 			count = 0;
 			foreach (var item in results)
 			{
-				Dictionary<char, int> c = new Dictionary<char, int>();
-				for (int x = 0; x < 5; x++)
+				Dictionary<char, long> c = new Dictionary<char, long>();
+				for (long x = 0; x < 5; x++)
 				{
 					//two of the same digits in a row
-					if (item.ToString()[x] == item.ToString()[x + 1])
+					if (item.ToString()[(int)x] == item.ToString()[(int)x + 1])
 					{
-						if (!c.ContainsKey(item.ToString()[x]))
-							c.Add(item.ToString()[x], 1);
+						if (!c.ContainsKey(item.ToString()[(int)x]))
+							c.Add(item.ToString()[(int)x], 1);
 						else
-							c[item.ToString()[x]]++;
+							c[item.ToString()[(int)x]]++;
 					}
 				}
