+++ AdventOfCode2019/Day9.cs
@@ -5,7 +5,7 @@ namespace AdventOfCode2019
 {
 	internal class Day9
 	{
-		internal static int Result1(string input)
+		internal static long Result1(string input)
 		{
 			string[] inputData = File.ReadAllLines(input);
 			inputData = inputData[0].Split(',');
@@ -15,9 +15,9 @@ namespace AdventOfCode2019
 			inputData = "1102,34915192,34915192,7,4,7,99,0".Split(',');
 			inputData = "104,1125899906842624,99".Split(',');
 
-			int[] data = new int[inputData.Length];
-			for (int i = 0; i < inputData.Length; i++)
-				data[i] = int.Parse(inputData[i]);
+			long[] data = new long[inputData.Length];
+			for (long i = 0; i < inputData.Length; i++)
+				data[i] = long.Parse(inputData[i]);
 
 			IntCode cpu = new IntCode(data);
 
@@ -27,7 +27,7 @@ namespace AdventOfCode2019
 			return cpu.output;
 		}
 
-		internal static int Result2(string input)
+		internal static long Result2(string input)
 		{
 			return 0;
 		}
